Flask_API/requirements.txt
fastapi
deta
uvicorn
python-dotenv